Parse #[ts(optional)] on structs
